Add Mishkat al-Masabih and improve Arabic tafsir pagination
Added Mishkat al-Masabih, a new hadith collection on sunnah.com. In addition, pagination for -atafsir now works such that clicking the backwards arrow will send the next page, while clicking the forwards arrow will send the previous page. This is because pages are flipped to the right when reading texts written in a RTL language in order.